Remove python2 from tests, OB3

+++ test/gninagrid/compare_bin.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 '''Compare two binmap files which are nothing but floating point numbers'''
 
@@ -14,7 +14,7 @@ buf1 = f1.read()
 buf2 = f2.read()
 
 assert len(buf1) == len(buf2)
-n = len(buf1)/4
+n = int(len(buf1)/4)
 
 vals1 = struct.unpack('f'*n,buf1)
 vals2 = struct.unpack('f'*n,buf2)
